+++ lib/alipay_payment.dart
@@ -4,6 +4,7 @@ library;
 export 'src/alipay_payment_method_channel.dart';
 export 'src/alipay_payment_platform_interface.dart';
 export 'src/alipay_payment_unsafe.dart';
+export 'src/alipay_payment_wait.dart';
 export 'src/models/alipay_auth_result.dart';
 export 'src/models/alipay_environment.dart';
 export 'src/models/alipay_result.dart';